+++ packages/x-randomvalues/src/react-native.ts
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 // Adapted from https://github.com/LinusU/react-native-get-random-values/blob/85f48393821c23b83b89a8177f56d3a81dc8b733/index.js
+//
 // Copyright (c) 2018, 2020 Linus Unnebäck
 // SPDX-License-Identifier: MIT
 
@@ -10,7 +11,7 @@ import { NativeModules } from 'react-native';
 import { base64Decode } from '@polkadot/wasm-util/base64';
 import { xglobal } from '@polkadot/x-global';
 
-import { getRandomValues as getRandomValuesBrowser } from './browser.js';
+import { crypto as cryptoBrowser, getRandomValues as getRandomValuesBrowser } from './browser.js';
 import { insecureRandomValues } from './fallback.js';
 
 export { packageInfo } from './packageInfo.js';
@@ -45,4 +46,10 @@ export const getRandomValues = (
     : (typeof xglobal.nativeCallSyncHook === 'undefined' || !NativeModules.ExpoRandom)
       ? insecureRandomValues
       : getRandomValuesRn
+);
+
+export const crypto = (
+  getRandomValues === getRandomValuesBrowser
+    ? cryptoBrowser
+    : { getRandomValues }
 );