+++ packages/x-randomvalues/src/shim.ts
@@ -1,0 +1,7 @@
+// Copyright 2017-2023 @polkadot/x-randomvalues authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+import { exposeGlobal } from '@polkadot/x-global';
+import { crypto } from '@polkadot/x-randomvalues';
+
+exposeGlobal('crypto', crypto);