+++ packages/x-textencoder/src/mod.ts
@@ -1,8 +1,4 @@
 // Copyright 2017-2023 @polkadot/x-textencoder authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-export { packageInfo } from './packageInfo.js';
-
-const _TextEncoder = TextEncoder;
-
-export { _TextEncoder as TextEncoder };
+export * from './browser.js';